[#noissue] Replace the static ServiceUidService utility with StaticServiceLookupService
The static-fallback lookup logic lived in two copies: an inline lambda
over the ServiceUidService static utility in
StaticServiceLookupConfiguration, and UidFetchers.DefaultUidFetcher.
Extract it into StaticServiceLookupService, a named
ServiceLookupService implementation in the collector, delegate both
callers to it, and delete ServiceUidService from commons-server.

Service lookup becomes mandatory in 4.0.0; remove the implementation
together with StaticServiceLookupConfiguration then.
