+++ collector/src/main/java/com/navercorp/pinpoint/io/request/UidFetchers.java
@@ -1,23 +1,14 @@
 package com.navercorp.pinpoint.io.request;
 
-import com.navercorp.pinpoint.common.server.uid.ServiceUid;
-import com.navercorp.pinpoint.common.server.uid.ServiceUidService;
-
-import java.util.concurrent.CompletableFuture;
+import com.navercorp.pinpoint.collector.uid.service.StaticServiceLookupService;
 
 public class UidFetchers {
 
-    public static final UidFetcher DEFAULT_UID_FETCHER = new DefaultUidFetcher();
+    private static final StaticServiceLookupService STATIC_SERVICE_LOOKUP_SERVICE = new StaticServiceLookupService();
+
+    public static final UidFetcher DEFAULT_UID_FETCHER = STATIC_SERVICE_LOOKUP_SERVICE::getServiceUid;
 
     public static UidFetcher defaultUidFetcher() {
         return DEFAULT_UID_FETCHER;
-    }
-
-    public static class DefaultUidFetcher implements UidFetcher {
-        @Override
-        public CompletableFuture<ServiceUid> getServiceUid(String serviceName) {
-            ServiceUid serviceUid = ServiceUidService.getServiceUid(serviceName);
-            return CompletableFuture.completedFuture(serviceUid);
-        }
     }
 }
