--- commons-server/src/main/java/com/navercorp/pinpoint/common/server/uid/ServiceUidService.java
@@ -1,18 +1,0 @@
-package com.navercorp.pinpoint.common.server.uid;
-
-import com.navercorp.pinpoint.common.util.StringUtils;
-
-public class ServiceUidService {
-
-
-    public static ServiceUid getServiceUid(String serviceName) {
-        if (StringUtils.isEmpty(serviceName)) {
-            return ServiceUid.DEFAULT;
-        }
-        if (ServiceUid.DEFAULT_SERVICE_UID_NAME.equals(serviceName)) {
-            return ServiceUid.DEFAULT;
-        }
-        // TODO ServiceUid query
-        return ServiceUid.DEFAULT;
-    }
-}
